+++ frontend/src/pages/board_page/waggle_page/WaggleDetailPage.jsx
@@ -1,5 +1,6 @@
 import React, { useEffect, useState } from "react";
 import { useParams, useNavigate, useLocation } from "react-router-dom";
+import Swal from "sweetalert2";
 import { useRecoilState } from "recoil";
 import styled from "styled-components";
 import {
@@ -112,22 +113,38 @@ function WaggleDetailPage() {
   };
 
   const handleDeleteBtn = () => {
-    deleteWaggle(
-      boardId,
-      () => {
-        if (state && state.fromMypage) {
-          navigate("/mypage", {
-            state: {
-              user: state.user,
-              category: state.category,
-            },
-          });
-        } else {
-          navigate("/waggle");
-        }
-      },
-      () => {},
-    );
+    Swal.fire({
+      html: `
+        <img src="https://raw.githubusercontent.com/Tarikul-Islam-Anik/Animated-Fluent-Emojis/master/Emojis/Animals/Bear.png" alt="Bear" width="100" height="100"/>
+        <p style='font-size:1.2rem; font-family:Pretendard-Bold;'>게시글을 삭제하시겠습니까?</p>
+      `,
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText:
+        "<div style='font-size:1rem; font-family:Pretendard-Regular;'>확인</div>",
+      cancelButtonText:
+        "<div style='font-size:1rem; font-family:Pretendard-Regular;'>취소</div>",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        deleteWaggle(
+          boardId,
+          () => {
+            if (state && state.fromMypage) {
+              navigate("/mypage", {
+                state: {
+                  user: state.user,
+                  category: state.category,
+                },
+              });
+            } else {
+              navigate("/waggle");
+            }
+          },
+          () => {},
+        );
+      }
+    });
   };
 
   const handleLikeClick = () => {
